1. Used the ChartType
2. Improved the bubble chart sample

+++ src/BlazorComponents/ChartJS/ChartJsBubbleDataset.cs
@@ -24,11 +24,11 @@
 
     public class BubbleData
     {
-        public int x { get; set; }
+        public double x { get; set; }
 
-        public int y { get; set; }
+        public double y { get; set; }
 
-        public int r { get; set; }
+        public double r { get; set; }
     }
 
     public enum BubbleChartPointStyle
